Move quantization code to separate file
